Add basic struct of message_2
As it is, there's a lot of duplication with the serialization of
message_1, meaning they could be combined. They're not, because this is
only the case due to not implementing all possible types of message_2.
Once the number of CBOR items in the sequence can change, we'll probably
need to have individual functions anyway.