Implement serialization/deserialization of COSE header map (ID_CRED_X)
We end up duplicating a lot of code. This could be solved better, maybe
by a trait with a default implementation based on the number of
elements.
But since this is possibly the last instance where I need to do this and
I'd like to replace all of this hacky CBOR stuff with a proper
implementation anyway, it'll stay like this for the moment.