Centralize code for transforming CBOR maps and arrays
Unfortunately this is still necessary. Rust's maps are statically typed,
so we can't mix types as is the case in most CBOR maps in the protocol.
The solution is the Value enum from serde_cbor, but this is not made
available in #![no_std], so we're stuck with this.

In cases like ID_CRED_x, where there's only a single key/value pair, we
could use maps directly, but since what we have works and its code
footprint isn't larger than the map solution, we'll stick with this in
these cases too.