Use architecture-dependent types where applicable
These shouldn't usually be used for anything other than memory sizes,
offsets and similar. But they are here, because we're representing very
small possible ranges of numbers (that fit in u8/i8), but it wouldn't
make sense to explicitly use these small types that need a full word
anyway and possibly a truncate instruction. The same goes for just using
u32/i32, because we might well run in a 16-bit embedded environment
where that would be wasteful too.