Check for error message before unwrapping from bstr
Because we don't have any guidance from the draft on whether the CBOR
sequence of an EDHOC error message is wrapped in a bstr for transmission
like the normal EDHOC messages, we assume that this is not the case.
This means we have to check before unwrapping, otherwise we'll fail
while trying to do so.