Flesh out the message trait
Still nowhere near good, but might be workable now. For getting the
options, the iterator solution would be nicer, although if we don't
want to mutate the options while we go over them (which would require
additional implementation on the user's part), we'll probably still
have to take the iterator, copy the options into a collection and then
modify them after the fact. Not much better than what we're doing now.