Remove CoAP message trait
The original plan was to use this to have an actual CoAP implementation
interact with messages for us, so we could be as CoAP agnostic as
possible. However, it turns out that the plaintext payload of our
messages is basically encoded as a CoAP message. Not completely, but as
a subset thereof. Most importantly, the options are encoded in the same
way. Of course it would be possible to write a CoAP options encoder and
decoder, but at that point, we might as well write full CoAP message
handling, at which point we don't need this complicated trait solution
anymore.
That's the reason why I instead copied and modified the necessary CoAP
handling from https://github.com/covertness/coap-rs, so this
implementation will be able to act directly on CoAP message bytes.